Fix response XML rejected by conforming parsers, validate output externally
Output was only ever checked by decoding it with the same encoding/xml that produced it, and Go's decoder tolerates duplicate attributes and matches elements by local name — so two defects survived the whole suite, the real fixtures, the soak test and CI. Every ItemValue with a ResultID declared its namespace prefix twice (invalid XML: expat, libxml2, .NET and JAXP all reject the document), and response payload elements went out in no namespace where the schema's elementFormDefault="qualified" requires the OPC XML-DA one. Separately, no array-typed Value could be constructed at all: backends could not report an ArrayOf<X> item, and standard property euInfo was unimplementable.

Also fixed: ItemValue's child order and DiagnosticInfo as an element per the schema sequence; a valueless property faulting a whole GetProperties/Browse response; the reaper killing a subscription mid-PolledRefresh; S_ codes dropping their value; deadband measured against the last read instead of the last reported value; ReturnAllItems discarding the buffer; poll-chain drift; xsd:int fields decoded as unsigned; UTC timestamps; plus new Config knobs (ErrorText, StatusCacheTTL, MaxBrowseElements, MaxTotalSubscribedItems) and a batch of smaller corrections. Each ships with a regression test confirmed to fail against the old behavior.

                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      all nine real third-party captures validating against it, and by seven
Verification no longer relies on the parser that wrote the output: server/wireformat_test.go asserts attribute uniqueness, namespace membership and schema child order in pure Go; testdata/golden/ pins the exact bytes of all eight operations plus a fault; testdata/schema/ holds the OPC XML-DA schema transcribed from the specification's WSDL, and a new CI job validates every generated response with xmllint. The transcription is corroborated by deliberate mutations each being rejected.

This changes the wire format — recorded responses and clients relying on the old element names, attribute-form DiagnosticInfo or Quality-before-Value order need regenerating.

+++ xmlda/itemparams_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestMergeItemParams_Precedence(t *testing.T) {
 	reqPath, listPath, itemPath := "req", "list", "item"
-	reqRate, itemRate := uint32(100), uint32(50)
+	reqRate, itemRate := int32(100), int32(50)
 
 	request := ItemParams{ItemPath: &reqPath, RequestedSamplingRate: &reqRate}
 	list := ItemParams{ItemPath: &listPath}
@@ -54,9 +54,9 @@ func TestMergeItemParams_EmptyStringItemPathIsMeaningful(t *testing.T) {
 
 func TestItemParamsAttrs_RoundTrip(t *testing.T) {
 	path := "my/item"
-	maxAge := int64(1000)
+	maxAge := int32(1000)
 	deadband := 5.5
-	rate := uint32(2000)
+	rate := int32(2000)
 	enable := true
 	reqType := QName{Space: XSDNamespace, Local: "int"}
